Update docker-compose.yml
Update docker compose to add a healthcheck to the db service.
The check pings sql for a while and considers the service unhealthy if the check fails to complete within 5 seconds or fails 3 times in a row.

Modified the backend service to not start unless the database is healthy

Now the backend will not fail due to starting before the db is ready